Add error message when auotmation's create rule REST call is faild. Move the JSONResponce from org.eclipse.smarthome.io.rest.core.internal to org.eclipse.smarthome.io.rest.core package.
